feat: add sysadmin check for channel delete

+++ src/channel/channel.manager.ts
@@ -52,34 +52,31 @@ export class ChannelManager {
         return ChannelRepository.updateMany(channelFilter, channel);
     }
 
-    static async deleteById(id: string, requestingUser: string) {
+    static async deleteById(id: string, requestingUser: string, isSystemAdmin: boolean) {
         let isPermitted: boolean = false;
         const currentChannel: IChannel | null = await ChannelManager.getById(id);
 
-        if (currentChannel) {
-            if (currentChannel.isProfile) throw new ProfileEditingIsForbiddenError();
+        if (!currentChannel) throw new ChannelNotFoundError();
+        if (currentChannel.isProfile) throw new ProfileEditingIsForbiddenError();
 
-            if (requestingUser === currentChannel.user) {
-                isPermitted = true;
-            } else {
-                const isAdmin: boolean = await UserPermissionsManager.isUserAdmin(requestingUser, id);
-                if (isAdmin) isPermitted = true;
-            }
-
-            if (isPermitted) {
-                const removed: IChannel | null = await ChannelRepository.deleteById(id);
-
-                if (removed) {
-                    ChannelBroker.publish('channelService.channel.remove.succeeded', { id });
-                }
-
-                return removed;
-            }
-
-            throw new UnauthorizedUserError();
+        if (isSystemAdmin) {
+            isPermitted = true;
+        } else if (requestingUser === currentChannel.user) {
+            isPermitted = true;
+        } else {
+            const isAdmin: boolean = await UserPermissionsManager.isUserAdmin(requestingUser, id);
+            if (isAdmin) isPermitted = true;
         }
 
-        throw new ChannelNotFoundError();
+        if (!isPermitted) throw new UnauthorizedUserError();
+
+        const removed: IChannel | null = await ChannelRepository.deleteById(id);
+
+        if (removed) {
+            ChannelBroker.publish('channelService.channel.remove.succeeded', { id });
+        }
+
+        return removed;
     }
 
     static getSearched(
